- Done Warm Up Exercises Details Screen ✅ - manage all states of warmup exercises get data - fix slider indecator not in the center

+++ lib/features/warm_up_exercises/data/models/warm_up_exercise_model.dart
@@ -4,7 +4,7 @@ class WarmUpExerciseModel {
   final String gifPath;
   final int? reps;
   final int? sets;
-  final int? weight;
+  final double? weight;
   final String? duration;
 
   WarmUpExerciseModel({
